cannon: Move state.go and state_mt.go into impls package
