+++ cannon/mipsevm/impls/multi_threaded/state_mt_test.go
@@ -1,14 +1,15 @@
-package mipsevm
+package multi_threaded
 
 import (
 	"debug/elf"
 	"encoding/json"
 	"testing"
 
-	"github.com/ethereum-optimism/optimism/cannon/mipsevm/core"
-	"github.com/ethereum-optimism/optimism/cannon/mipsevm/patch"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/stretchr/testify/require"
+
+	"github.com/ethereum-optimism/optimism/cannon/mipsevm/core"
+	"github.com/ethereum-optimism/optimism/cannon/mipsevm/patch"
 )
 
 func setWitnessField(witness MTStateWitness, fieldOffset int, fieldData []byte) {
@@ -83,7 +84,7 @@ func TestMTState_EncodeWitness(t *testing.T) {
 }
 
 func TestMTState_JSONCodec(t *testing.T) {
-	elfProgram, err := elf.Open("../example/bin/hello.elf")
+	elfProgram, err := elf.Open("../../../example/bin/hello.elf")
 	require.NoError(t, err, "open ELF file")
 	state, err := patch.LoadELF(elfProgram, CreateInitialMTState)
 	require.NoError(t, err, "load ELF into state")
