+++ cannon/mipsevm/impls/single_threaded/state.go
@@ -1,14 +1,15 @@
-package mipsevm
+package single_threaded
 
 import (
 	"encoding/binary"
 	"encoding/json"
 	"fmt"
 
-	"github.com/ethereum-optimism/optimism/cannon/mipsevm/core"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/crypto"
+
+	"github.com/ethereum-optimism/optimism/cannon/mipsevm/core"
 )
 
 // STATE_WITNESS_SIZE is the size of the state witness encoding in bytes.
