+++ cannon/mipsevm/tests/fuzz_evm_test.go
@@ -6,15 +6,16 @@ import (
 	"os"
 	"testing"
 
-	"github.com/ethereum-optimism/optimism/cannon/mipsevm"
-	"github.com/ethereum-optimism/optimism/cannon/mipsevm/core"
-	"github.com/ethereum-optimism/optimism/cannon/mipsevm/test_util"
-
-	preimage "github.com/ethereum-optimism/optimism/op-preimage"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/stretchr/testify/require"
+
+	"github.com/ethereum-optimism/optimism/cannon/mipsevm"
+	"github.com/ethereum-optimism/optimism/cannon/mipsevm/core"
+	"github.com/ethereum-optimism/optimism/cannon/mipsevm/impls/single_threaded"
+	"github.com/ethereum-optimism/optimism/cannon/mipsevm/test_util"
+	preimage "github.com/ethereum-optimism/optimism/op-preimage"
 )
 
 const syscallInsn = uint32(0x00_00_00_0c)
@@ -24,7 +25,7 @@ func FuzzStateSyscallBrk(f *testing.F) {
 	f.Fuzz(func(t *testing.T, pc uint32, step uint64, preimageOffset uint32) {
 		pc = pc & 0xFF_FF_FF_FC // align PC
 		nextPC := pc + 4
-		state := &mipsevm.State{
+		state := &single_threaded.State{
 			Cpu: core.CpuScalars{
 				PC:     pc,
 				NextPC: nextPC,
@@ -64,7 +65,7 @@ func FuzzStateSyscallBrk(f *testing.F) {
 		require.Equal(t, preimageOffset, state.PreimageOffset)
 
 		evm := test_util.NewMIPSEVM(contracts, addrs)
-		evmPost := evm.Step(t, stepWitness, step, mipsevm.GetStateHashFn())
+		evmPost := evm.Step(t, stepWitness, step, single_threaded.GetStateHashFn())
 		goPost, _ := goState.GetState().EncodeWitness()
 		require.Equal(t, hexutil.Bytes(goPost).String(), hexutil.Bytes(evmPost).String(),
 			"mipsevm produced different state than EVM")
@@ -76,7 +77,7 @@ func FuzzStateSyscallClone(f *testing.F) {
 	f.Fuzz(func(t *testing.T, pc uint32, step uint64, preimageOffset uint32) {
 		pc = pc & 0xFF_FF_FF_FC // align PC
 		nextPC := pc + 4
-		state := &mipsevm.State{
+		state := &single_threaded.State{
 			Cpu: core.CpuScalars{
 				PC:     pc,
 				NextPC: nextPC,
@@ -115,7 +116,7 @@ func FuzzStateSyscallClone(f *testing.F) {
 		require.Equal(t, preimageOffset, state.PreimageOffset)
 
 		evm := test_util.NewMIPSEVM(contracts, addrs)
-		evmPost := evm.Step(t, stepWitness, step, mipsevm.GetStateHashFn())
+		evmPost := evm.Step(t, stepWitness, step, single_threaded.GetStateHashFn())
 		goPost, _ := goState.GetState().EncodeWitness()
 		require.Equal(t, hexutil.Bytes(goPost).String(), hexutil.Bytes(evmPost).String(),
 			"mipsevm produced different state than EVM")
@@ -126,7 +127,7 @@ func FuzzStateSyscallMmap(f *testing.F) {
 	contracts, addrs := testContractsSetup(f)
 	step := uint64(0)
 	f.Fuzz(func(t *testing.T, addr uint32, siz uint32, heap uint32) {
-		state := &mipsevm.State{
+		state := &single_threaded.State{
 			Cpu: core.CpuScalars{
 				PC:     0,
 				NextPC: 4,
@@ -177,7 +178,7 @@ func FuzzStateSyscallMmap(f *testing.F) {
 		}
 
 		evm := test_util.NewMIPSEVM(contracts, addrs)
-		evmPost := evm.Step(t, stepWitness, step, mipsevm.GetStateHashFn())
+		evmPost := evm.Step(t, stepWitness, step, single_threaded.GetStateHashFn())
 		goPost, _ := goState.GetState().EncodeWitness()
 		require.Equal(t, hexutil.Bytes(goPost).String(), hexutil.Bytes(evmPost).String(),
 			"mipsevm produced different state than EVM")
@@ -189,7 +190,7 @@ func FuzzStateSyscallExitGroup(f *testing.F) {
 	f.Fuzz(func(t *testing.T, exitCode uint8, pc uint32, step uint64) {
 		pc = pc & 0xFF_FF_FF_FC // align PC
 		nextPC := pc + 4
-		state := &mipsevm.State{
+		state := &single_threaded.State{
 			Cpu: core.CpuScalars{
 				PC:     pc,
 				NextPC: nextPC,
@@ -227,7 +228,7 @@ func FuzzStateSyscallExitGroup(f *testing.F) {
 		require.Equal(t, uint32(0), state.PreimageOffset)
 
 		evm := test_util.NewMIPSEVM(contracts, addrs)
-		evmPost := evm.Step(t, stepWitness, step, mipsevm.GetStateHashFn())
+		evmPost := evm.Step(t, stepWitness, step, single_threaded.GetStateHashFn())
 		goPost, _ := goState.GetState().EncodeWitness()
 		require.Equal(t, hexutil.Bytes(goPost).String(), hexutil.Bytes(evmPost).String(),
 			"mipsevm produced different state than EVM")
@@ -238,7 +239,7 @@ func FuzzStateSyscallFcntl(f *testing.F) {
 	contracts, addrs := testContractsSetup(f)
 	step := uint64(0)
 	f.Fuzz(func(t *testing.T, fd uint32, cmd uint32) {
-		state := &mipsevm.State{
+		state := &single_threaded.State{
 			Cpu: core.CpuScalars{
 				PC:     0,
 				NextPC: 4,
@@ -293,7 +294,7 @@ func FuzzStateSyscallFcntl(f *testing.F) {
 		}
 
 		evm := test_util.NewMIPSEVM(contracts, addrs)
-		evmPost := evm.Step(t, stepWitness, step, mipsevm.GetStateHashFn())
+		evmPost := evm.Step(t, stepWitness, step, single_threaded.GetStateHashFn())
 		goPost, _ := goState.GetState().EncodeWitness()
 		require.Equal(t, hexutil.Bytes(goPost).String(), hexutil.Bytes(evmPost).String(),
 			"mipsevm produced different state than EVM")
@@ -305,7 +306,7 @@ func FuzzStateHintRead(f *testing.F) {
 	step := uint64(0)
 	f.Fuzz(func(t *testing.T, addr uint32, count uint32) {
 		preimageData := []byte("hello world")
-		state := &mipsevm.State{
+		state := &single_threaded.State{
 			Cpu: core.CpuScalars{
 				PC:     0,
 				NextPC: 4,
@@ -346,7 +347,7 @@ func FuzzStateHintRead(f *testing.F) {
 		require.Equal(t, expectedRegisters, state.Registers)
 
 		evm := test_util.NewMIPSEVM(contracts, addrs)
-		evmPost := evm.Step(t, stepWitness, step, mipsevm.GetStateHashFn())
+		evmPost := evm.Step(t, stepWitness, step, single_threaded.GetStateHashFn())
 		goPost, _ := goState.GetState().EncodeWitness()
 		require.Equal(t, hexutil.Bytes(goPost).String(), hexutil.Bytes(evmPost).String(),
 			"mipsevm produced different state than EVM")
@@ -361,7 +362,7 @@ func FuzzStatePreimageRead(f *testing.F) {
 		if preimageOffset >= uint32(len(preimageData)) {
 			t.SkipNow()
 		}
-		state := &mipsevm.State{
+		state := &single_threaded.State{
 			Cpu: core.CpuScalars{
 				PC:     0,
 				NextPC: 4,
@@ -413,7 +414,7 @@ func FuzzStatePreimageRead(f *testing.F) {
 		require.Equal(t, preStatePreimageKey, state.PreimageKey)
 
 		evm := test_util.NewMIPSEVM(contracts, addrs)
-		evmPost := evm.Step(t, stepWitness, step, mipsevm.GetStateHashFn())
+		evmPost := evm.Step(t, stepWitness, step, single_threaded.GetStateHashFn())
 		goPost, _ := goState.GetState().EncodeWitness()
 		require.Equal(t, hexutil.Bytes(goPost).String(), hexutil.Bytes(evmPost).String(),
 			"mipsevm produced different state than EVM")
@@ -425,7 +426,7 @@ func FuzzStateHintWrite(f *testing.F) {
 	step := uint64(0)
 	f.Fuzz(func(t *testing.T, addr uint32, count uint32, randSeed int64) {
 		preimageData := []byte("hello world")
-		state := &mipsevm.State{
+		state := &single_threaded.State{
 			Cpu: core.CpuScalars{
 				PC:     0,
 				NextPC: 4,
@@ -474,7 +475,7 @@ func FuzzStateHintWrite(f *testing.F) {
 		require.Equal(t, expectedRegisters, state.Registers)
 
 		evm := test_util.NewMIPSEVM(contracts, addrs)
-		evmPost := evm.Step(t, stepWitness, step, mipsevm.GetStateHashFn())
+		evmPost := evm.Step(t, stepWitness, step, single_threaded.GetStateHashFn())
 		goPost, _ := goState.GetState().EncodeWitness()
 		require.Equal(t, hexutil.Bytes(goPost).String(), hexutil.Bytes(evmPost).String(),
 			"mipsevm produced different state than EVM")
@@ -486,7 +487,7 @@ func FuzzStatePreimageWrite(f *testing.F) {
 	step := uint64(0)
 	f.Fuzz(func(t *testing.T, addr uint32, count uint32) {
 		preimageData := []byte("hello world")
-		state := &mipsevm.State{
+		state := &single_threaded.State{
 			Cpu: core.CpuScalars{
 				PC:     0,
 				NextPC: 4,
@@ -530,7 +531,7 @@ func FuzzStatePreimageWrite(f *testing.F) {
 		require.Equal(t, expectedRegisters, state.Registers)
 
 		evm := test_util.NewMIPSEVM(contracts, addrs)
-		evmPost := evm.Step(t, stepWitness, step, mipsevm.GetStateHashFn())
+		evmPost := evm.Step(t, stepWitness, step, single_threaded.GetStateHashFn())
 		goPost, _ := goState.GetState().EncodeWitness()
 		require.Equal(t, hexutil.Bytes(goPost).String(), hexutil.Bytes(evmPost).String(),
 			"mipsevm produced different state than EVM")
