+++ op-challenger/game/fault/trace/cannon/provider_test.go
@@ -11,16 +11,17 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/ethereum-optimism/optimism/cannon/mipsevm"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/log"
+	"github.com/stretchr/testify/require"
+
 	"github.com/ethereum-optimism/optimism/cannon/mipsevm/core"
+	"github.com/ethereum-optimism/optimism/cannon/mipsevm/impls/single_threaded"
 	"github.com/ethereum-optimism/optimism/op-challenger/game/fault/trace/utils"
 	"github.com/ethereum-optimism/optimism/op-challenger/game/fault/trace/vm"
 	"github.com/ethereum-optimism/optimism/op-challenger/game/fault/types"
 	"github.com/ethereum-optimism/optimism/op-service/ioutil"
 	"github.com/ethereum-optimism/optimism/op-service/testlog"
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/log"
-	"github.com/stretchr/testify/require"
 )
 
 //go:embed test_data
@@ -50,7 +51,7 @@ func TestGet(t *testing.T) {
 
 	t.Run("ProofAfterEndOfTrace", func(t *testing.T) {
 		provider, generator := setupWithTestData(t, dataDir, prestate)
-		generator.finalState = &mipsevm.State{
+		generator.finalState = &single_threaded.State{
 			Memory: &core.Memory{},
 			Step:   10,
 			Exited: true,
@@ -106,7 +107,7 @@ func TestGetStepData(t *testing.T) {
 	t.Run("GenerateProof", func(t *testing.T) {
 		dataDir, prestate := setupTestData(t)
 		provider, generator := setupWithTestData(t, dataDir, prestate)
-		generator.finalState = &mipsevm.State{
+		generator.finalState = &single_threaded.State{
 			Memory: &core.Memory{},
 			Step:   10,
 			Exited: true,
@@ -132,7 +133,7 @@ func TestGetStepData(t *testing.T) {
 	t.Run("ProofAfterEndOfTrace", func(t *testing.T) {
 		dataDir, prestate := setupTestData(t)
 		provider, generator := setupWithTestData(t, dataDir, prestate)
-		generator.finalState = &mipsevm.State{
+		generator.finalState = &single_threaded.State{
 			Memory: &core.Memory{},
 			Step:   10,
 			Exited: true,
@@ -158,7 +159,7 @@ func TestGetStepData(t *testing.T) {
 	t.Run("ReadLastStepFromDisk", func(t *testing.T) {
 		dataDir, prestate := setupTestData(t)
 		provider, initGenerator := setupWithTestData(t, dataDir, prestate)
-		initGenerator.finalState = &mipsevm.State{
+		initGenerator.finalState = &single_threaded.State{
 			Memory: &core.Memory{},
 			Step:   10,
 			Exited: true,
@@ -176,7 +177,7 @@ func TestGetStepData(t *testing.T) {
 		require.Contains(t, initGenerator.generated, 7000, "should have tried to generate the proof")
 
 		provider, generator := setupWithTestData(t, dataDir, prestate)
-		generator.finalState = &mipsevm.State{
+		generator.finalState = &single_threaded.State{
 			Memory: &core.Memory{},
 			Step:   10,
 			Exited: true,
@@ -248,7 +249,7 @@ func setupWithTestData(t *testing.T, dataDir string, prestate string) (*CannonTr
 
 type stubGenerator struct {
 	generated  []int // Using int makes assertions easier
-	finalState *mipsevm.State
+	finalState *single_threaded.State
 	proof      *utils.ProofData
 }
