+++ op-challenger/game/fault/trace/cannon/state.go
@@ -5,11 +5,11 @@ import (
 	"fmt"
 	"io"
 
-	"github.com/ethereum-optimism/optimism/cannon/mipsevm"
+	"github.com/ethereum-optimism/optimism/cannon/mipsevm/impls/single_threaded"
 	"github.com/ethereum-optimism/optimism/op-service/ioutil"
 )
 
-func parseState(path string) (*mipsevm.State, error) {
+func parseState(path string) (*single_threaded.State, error) {
 	file, err := ioutil.OpenDecompressed(path)
 	if err != nil {
 		return nil, fmt.Errorf("cannot open state file (%v): %w", path, err)
@@ -17,9 +17,9 @@ func parseState(path string) (*mipsevm.State, error) {
 	return parseStateFromReader(file)
 }
 
-func parseStateFromReader(in io.ReadCloser) (*mipsevm.State, error) {
+func parseStateFromReader(in io.ReadCloser) (*single_threaded.State, error) {
 	defer in.Close()
-	var state mipsevm.State
+	var state single_threaded.State
 	if err := json.NewDecoder(in).Decode(&state); err != nil {
 		return nil, fmt.Errorf("invalid mipsevm state: %w", err)
 	}
